+++ doc/conf.py
@@ -24,9 +24,13 @@ import os
 sys.path.insert(0, os.path.abspath('..'))
 
 # -- General configuration ------------------------------------------------
+on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
 
 # If your documentation needs a minimal Sphinx version, state it here.
-needs_sphinx = '1.6.9'
+if on_rtd:
+    needs_sphinx = '1.6.9' # this one is needed by RTD
+else:
+    needs_sphinx = '1.4.9' # this one available at CNRM
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -117,7 +121,6 @@ pygments_style = 'sphinx'
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 html_theme = 'default'
-on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
 if on_rtd:
     rtd_theme=False
     if rtd_theme :